continuing to restructure tests
