+++ test/logic/checkPivotCategories.js
@@ -1,0 +1,20 @@
+import chai from 'chai';
+import { checkPivotCategories } from '../../src/logic.js';
+
+chai.expect();
+const expect = chai.expect;
+
+export default () => {
+  const actualCategories = ['pig', 'fish', 'dog'];
+
+  it('should return undefined if pivot category exists', () => {
+    expect(checkPivotCategories(actualCategories, ['dog'])).to.be.undefined;
+  });
+
+  it('should throw an error if pivot category does not exist', () => {
+    const errorFunc = checkPivotCategories.bind(
+        null, actualCategories, ['cat']);
+
+    expect(errorFunc).to.throw(Error);
+  });
+};
